define function to send a post request to the database update the recipes table with api id and recipe title. Found bug where sign-up window hash update string is invalud. WIP. #13
