+++ client/lib/index.js
@@ -1,3 +1,4 @@
 export { default as parseRoute } from './parse-route';
 export { default as arrayToString } from './array-to-string';
 export { default as searchString } from './search-string';
+export { default as dbPostRecipe } from './db-post-recipe';
